requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-mongoengine==1.0.0
flask-restx==1.3.0
Flask-WTF==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mcp==1.11.0
mongoengine==0.29.1
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pluggy==1.5.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.11.3
pyrsistent==0.20.0
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
